Move getAllApprovedPremisesApplications to Cas1ApplicationService.
